reorganize fedex tests by features
